remove type organization
